Add $VENDORDIR to load path to fix linking and remove libsodium from download list for now to temporarily defer fixing android deploy wrinkle:
libzmq.so is depending on libsodium.so.4, but libsodium is being copied under the name libsodium.so, and the android linker isn't recognizing it as the same.  furthermore, ANDROID_EXTRA_LIBS is rejecting any names that don't end with *.so, so it refuses to bundle in libsodium if it is named libsodium.so.4 to satisfy the libzmq.so dependency.  Solving this problem (and consequently, supporting the libzmq v4 encryption features) is being deferred to a later date, and this description of the problem should help at that time.